+++ x/finality/keeper/msg_server_test.go
@@ -555,6 +555,7 @@ func TestBtcDelegationRewards(t *testing.T) {
 	// process the events of the activated BTC delegations
 	h.BTCStakingKeeper.IndexBTCHeight(h.Ctx)
 	h.FinalityKeeper.UpdatePowerDist(h.Ctx)
+	h.IncentivesKeeper.ProcessRewardTrackerEventsAtHeight(h.Ctx, uint64(h.Ctx.HeaderInfo().Height))
 
 	fp1CurrentRwd, err := h.IncentivesKeeper.GetFinalityProviderCurrentRewards(h.Ctx, fp1.Address())
 	h.NoError(err)
@@ -627,6 +628,7 @@ func TestBtcDelegationRewardsEarlyUnbondingAndExpire(t *testing.T) {
 	// process the events as active btc delegation
 	h.BTCStakingKeeper.IndexBTCHeight(h.Ctx)
 	h.FinalityKeeper.UpdatePowerDist(h.Ctx)
+	h.IncentivesKeeper.ProcessRewardTrackerEventsAtHeight(h.Ctx, uint64(h.Ctx.HeaderInfo().Height))
 
 	h.EqualBtcDelRwdTrackerActiveSat(fp.Address(), del.Address(), uint64(stakingValue))
 
@@ -662,6 +664,7 @@ func TestBtcDelegationRewardsEarlyUnbondingAndExpire(t *testing.T) {
 	// process the events as early unbonding btc delegation
 	h.BTCStakingKeeper.IndexBTCHeight(h.Ctx)
 	h.FinalityKeeper.UpdatePowerDist(h.Ctx)
+	h.IncentivesKeeper.ProcessRewardTrackerEventsAtHeight(h.Ctx, uint64(h.Ctx.HeaderInfo().Height))
 
 	h.EqualBtcDelRwdTrackerActiveSat(fp.Address(), del.Address(), 0)
 
@@ -676,6 +679,7 @@ func TestBtcDelegationRewardsEarlyUnbondingAndExpire(t *testing.T) {
 	// process the events as expired btc delegation
 	h.BTCStakingKeeper.IndexBTCHeight(h.Ctx)
 	h.FinalityKeeper.UpdatePowerDist(h.Ctx)
+	h.IncentivesKeeper.ProcessRewardTrackerEventsAtHeight(h.Ctx, uint64(h.Ctx.HeaderInfo().Height))
 
 	h.EqualBtcDelRwdTrackerActiveSat(fp.Address(), del.Address(), 0)
 }